Move API to CustomDialogService
