+++ extensions/mssql/src/errorDiagnostics/errorDiagnosticsProvider.ts
@@ -83,12 +83,13 @@ export class ErrorDiagnosticsProvider extends SqlOpsFeature<any> {
 					};
 
 					actions.push(trustServerCertAction);
-					const result = await azdata.connection.openCustomErrorDialog(
+					const result = await azdata.window.openCustomErrorDialog(
 						{
 							severity: azdata.window.MessageLevel.Error,
 							headerTitle: ErrorDiagnosticsConstants.ConnectionErrorDialogTitle,
 							message: errorMessage,
 							messageDetails: callStack,
+							telemetryView: 'CustomizedConnectionErrorDialog',
 							instructionText: ErrorDiagnosticsConstants.TSC_InstructionText,
 							readMoreLink: ErrorDiagnosticsConstants.TSC_ReadMoreLink,
 							actions: actions
