+++ src/sql/azdata.proposed.d.ts
@@ -537,14 +537,6 @@ declare module 'azdata' {
 		 * @returns The new password that is returned from the operation or undefined if unsuccessful.
 		 */
 		export function openChangePasswordDialog(profile: IConnectionProfile): Thenable<string | undefined>;
-
-		/**
-		 * Opens the error dialog with customization options provided.
-		 * @param options Dialog options to customize error dialog.
-		 * @returns Id of action button clicked by user, e.g. ok, cancel
-		 */
-		export function openCustomErrorDialog(options: window.IErrorDialogOptions): Thenable<string | undefined>;
-
 	}
 
 	/*
@@ -1831,6 +1823,13 @@ declare module 'azdata' {
 		}
 
 		/**
+		 * Opens the error dialog with customization options provided.
+		 * @param options Dialog options to customize error dialog.
+		 * @returns Id of action button clicked by user, e.g. ok, cancel
+		 */
+		export function openCustomErrorDialog(options: IErrorDialogOptions): Thenable<string | undefined>;
+
+		/**
 		 * Provides dialog options to customize modal dialog content and layout
 		 */
 		export interface IErrorDialogOptions {
@@ -1850,6 +1849,10 @@ declare module 'azdata' {
 			 * (Optional) Detailed message, e.g stack trace of error.
 			 */
 			messageDetails?: string;
+			/**
+			 * Telemetry View to be used for emitting telemetry events.
+			 */
+			telemetryView?: string,
 			/**
 			 * (Optional) List of custom actions to include in modal dialog alongwith a 'Cancel' button.
 			 * If custom 'actions' are not provided, 'OK' button will be shown by default.
