+++ src/sql/workbench/api/common/extHostConnectionManagement.ts
@@ -9,7 +9,6 @@ import * as azdata from 'azdata';
 import { IDisposable } from 'vs/base/common/lifecycle';
 import { Disposable } from 'vs/workbench/api/common/extHostTypes';
 import { SqlMainContext } from 'vs/workbench/api/common/extHost.protocol';
-import { IErrorDialogOptions } from 'sql/workbench/api/common/sqlExtHostTypes';
 
 export class ExtHostConnectionManagement extends ExtHostConnectionManagementShape {
 
@@ -77,10 +76,6 @@ export class ExtHostConnectionManagement extends ExtHostConnectionManagementShap
 
 	public $openChangePasswordDialog(profile: azdata.IConnectionProfile): Thenable<string | undefined> {
 		return this._proxy.$openChangePasswordDialog(profile);
-	}
-
-	public $openCustomErrorDialog(options: IErrorDialogOptions): Thenable<string | undefined> {
-		return this._proxy.$openCustomErrorDialog(options);
 	}
 
 	public $listDatabases(connectionId: string): Thenable<string[]> {
