+++ src/sql/workbench/services/dialog/browser/customDialogService.ts
@@ -8,6 +8,8 @@ import { WizardModal } from 'sql/workbench/services/dialog/browser/wizardModal';
 import { Dialog, Wizard } from 'sql/workbench/services/dialog/common/dialogTypes';
 import { IModalOptions } from 'sql/workbench/browser/modal/modal';
 import { IInstantiationService } from 'vs/platform/instantiation/common/instantiation';
+import { ErrorMessageDialog } from 'sql/workbench/services/errorMessage/browser/errorMessageDialog';
+import { IErrorDialogOptions } from 'sql/workbench/api/common/sqlExtHostTypes';
 
 export const DefaultDialogOptions: IModalOptions = { hasBackButton: false, width: 'narrow', hasErrors: true, hasSpinner: true };
 export const DefaultWizardOptions: IModalOptions = { hasBackButton: false, width: 'wide', hasErrors: true, hasSpinner: true };
@@ -50,5 +52,17 @@ export class CustomDialogService {
 
 	public getWizardModal(wizard: Wizard): WizardModal | undefined {
 		return this._wizardModals.get(wizard);
+	}
+
+	/**
+	 * Shows error dialog customized with given options
+	 * @param options Error Dialog options to customize error message dialog.
+	 * @param telemetryView Telemetry View to be used for dispatching telemetry events.
+	 */
+	public async openCustomErrorDialog(options: IErrorDialogOptions): Promise<string | undefined> {
+		let dialog = this._instantiationService.createInstance(ErrorMessageDialog);
+		dialog.render();
+		let result = await dialog.openCustomAsync(options);
+		return result;
 	}
 }
