+++ src/sql/workbench/services/errorMessage/browser/errorMessageService.ts
@@ -11,7 +11,6 @@ import { IAction } from 'vs/base/common/actions';
 import { ErrorMessageDialog } from 'sql/workbench/services/errorMessage/browser/errorMessageDialog';
 import { IErrorMessageService } from 'sql/platform/errorMessage/common/errorMessageService';
 import { TelemetryView } from 'sql/platform/telemetry/common/telemetryKeys';
-import { IErrorDialogOptions, MessageLevel } from 'sql/workbench/api/common/sqlExtHostTypes';
 
 export class ErrorMessageService implements IErrorMessageService {
 
@@ -41,13 +40,6 @@ export class ErrorMessageService implements IErrorMessageService {
 		return this._errorDialog.open(telemetryView, severity, title, message, messageDetails, actions, instructionText, readMoreLink);
 	}
 
-	public async showDialogAsync(options: IErrorDialogOptions): Promise<string | undefined> {
-		let dialog = this._instantiationService.createInstance(ErrorMessageDialog);
-		dialog.render();
-		let result = await dialog.openCustomAsync(TelemetryView.ConnectionErrorDialog, this.convertToSeverity(options.severity), options);
-		return result;
-	}
-
 	private getDefaultTitle(severity: Severity) {
 		switch (severity) {
 			case Severity.Error:
@@ -59,21 +51,5 @@ export class ErrorMessageService implements IErrorMessageService {
 			case Severity.Ignore:
 				return localize('ignore', "Ignore");
 		}
-	}
-
-	private convertToSeverity(messageLevel: MessageLevel): Severity {
-		let severity: Severity = Severity.Error;
-		switch (messageLevel) {
-			case MessageLevel.Error:
-				severity = Severity.Error;
-				break;
-			case MessageLevel.Information:
-				severity = Severity.Info;
-				break;
-			case MessageLevel.Warning:
-				severity = Severity.Warning;
-				break;
-		}
-		return severity;
 	}
 }
